local_parser/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
beautifulsoup4==4.12.3
